Upgrade Spring to latest version. Adjust Swagger accordingly

+++ src/main/java/by/andd3dfx/templateapp/services/impl/ArticleService.java
@@ -54,11 +54,9 @@ public class ArticleService implements IArticleService {
     @Transactional
     @Override
     public void delete(Long id) {
-        try {
+        if (articleRepository.existsById(id)) {
             articleRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException ex) {
-            throw new ArticleNotFoundException(id);
-        }
+        } else throw new ArticleNotFoundException(id);
     }
 
     @Transactional(readOnly = true)
